Market file
In this file, all offers and requests that have been made by players, will be stored. It is dynamic in that the order of the file is chronological; the offers and requests on top have been made first. On the market, that's what gets priority; the first to come, will be the first to get the deal.